fixed non-negative likelihood due to (a) insufficient precision of float16 and (b) bad handling of models with zero total coverage

+++ mglex/types.py
@@ -6,7 +6,7 @@ Submodule with all type definitions.
 
 import numpy as np
 
-prob_type = np.float16
+prob_type = np.float32
 logprob_type = np.float16
 large_float_type = np.float32
 seqlen_type = np.uint32
